Rectify extraction of the enum definitions
To avoid complications in parsing the enum definitions with pycparser the header gets
preprocessed and thus the extract_enums() function must yield only the typedefs belonging to
the specific header, not all #included ones.

Also, collect the position of each element and add a docstring with a reference to the source
on Github.