Write documentation of extracted enums into reST documents
This is mainly to workaround to the fact that currently the documentation on readthedocs.org is
built with an old Sphinx run by Python 2, and thus it's not able to parse the enums
subpackages.

On the other hand, it is of little value to put such a simple docstring (it's just an
hyperlink) on each class.