First step into using correct positions within the original Unicode string
The C functions exposed by the underlying libpg_query operates on the UTF-8 encoded C string,
and thus both the nodes "location" and the possible errors position are relative to that, not
to the original Python string: when it contains non-ASCII, the offsets are clearly different.

This fixes the issue with the exceptions raised in case of errors and with the tokens info
returned by the new scan() function, but more work is needed to adjust the "location" slot of
the AST nodes: we need that to make further steps of issue #23.