Drop precomputed FETCH_ALL constant
In the PG headers FETCH_ALL is a synonym of LONG_MAX, and the latter
value depends on the actual architecture so cannot be precomputed and
hardcoded in the extracted enums.

Given that it's used in a single place, instead of doing fancy gymnic to
compute it lazily, take the straight path and directly use the LONG_MAX
symbol exposed in the parser module, that correspond to its homonym C
value.

This fixes issue #142.

While there, tweak the FetchStmt printer to emit slightly nicer
representation.