Support Windows
Compiling on Windows is always a pain (IMHO, of course :-), but with a
sufficient amount of patience I was eventually able to do it, first on
my local machine and then tweaking the CI configuration to build wheels
on GitHub.

Due to the way PG's headers has been extended to support Windows, its
main header (that is, postgres.h) must be included earlier than any
other system header, and that means before Python.h.

Since I could not find a way to convince Cython to generate a C source
that includes a particular header before anything else, I replaced the
`cythonize` approach with an explicit pre-generation of the C source
using the `cython` CLI tool and the hacking it with `ed` to move the
import of "postgres.h" near the beginning.

This fixes issue #7.