Replace invisible \feff character in expected result with placeholder
Recently added test for issue #170 failed in the CI Windows suite, for
unknown reasons: apparently the invisible ZERO WIDTH NO-BREAK SPACE
character embedded in the "expected" value disappeared, once loaded in
the Windows environment.

This replaces that glyph with its Unicode name, and the test swaps it
with the real glyph at execution time, hopefully repairing the case on
all the architectures, let's see...

In any case, it may be preferred as it makes more evident the underlying
trick.